refactor: split god modules, wire run_print to agent_loop, dedup adapters, shore up test coverage
Wave 1 — Split god modules (-2,329 lines):
- permission/checker.rs → engine.rs + path.rs + allowlist.rs
- session/mod.rs → compact.rs
- provider.rs → provider/ (client.rs + summarize.rs)
- plugin/mod.rs → loader.rs
- ui/renderer.rs → panel_data.rs + buffer.rs
- ui/mod.rs → permission_ui.rs + tool_display.rs (-1,358, -20%)

Wave 2 — Migration cleanup (-302 lines):
- Wire AnyAgent::run_print through spawn_runner/agent_loop
- Delete old MultiTurnStream path from runner.rs (-248 lines)
- Remove #[allow(dead_code)] from ui/tui/mod.rs
- Dedup 18 thin wrapper methods across 9 semantic adapters

Wave 3 — Test coverage:
- +9 session persistence integration tests (round-trip messages,
  tool calls, permission allowlist, compactions, tree/store,
  plugin entries, re-save metadata, delete, branch summaries)
- Verified 52 permission tests, 24 provider tests, 241 agent_loop tests

Wave 4 — CI gate:
- cargo fmt passes
- cargo clippy --all-features passes (174 pre-existing warnings)
- cargo test --all-features: 1456 pass, 1 pre-existing failure
  (unrelated Clojure tree-sitter test)

Total: -2,631 lines, 12 new files, zero new bugs.
