+++ src/ui/mod.rs
@@ -1,11 +1,14 @@
 pub(crate) mod ansi;
 pub(crate) mod avatar;
 pub(crate) mod box_render;
+pub(crate) mod buffer;
 mod events;
 mod highlight;
 pub(crate) mod input;
 mod markdown;
 pub(crate) mod notifications;
+pub(crate) mod panel_data;
+pub(crate) mod permission_ui;
 pub(crate) mod picker;
 #[cfg(feature = "plugin")]
 mod plugin_tree;
@@ -18,6 +21,7 @@ mod streaming;
 pub(crate) mod sysload;
 pub(crate) mod terminal;
 pub(crate) mod theme;
+pub(crate) mod tool_display;
 mod tree;
 /// ui-redesign: ratatui-based render pipeline. Lives alongside the
 /// legacy `renderer` module during the staged migration; see beads
@@ -60,6 +64,7 @@ use crate::ui::renderer::{LineEntry, Renderer};
 use crate::ui::slash::{handle_compress, handle_slash};
 use crate::ui::status::StatusLine;
 use crate::ui::terminal::TerminalGuard;
+use tool_display::*;
 
 // Themed color accessors. These wrap `theme::agent()` etc. so we can
 // keep the existing call-site spelling (e.g. `c_agent()` is now a fn).
@@ -618,187 +623,6 @@ fn sanitize_single_line(s: &str, max_chars: usize) -> String {
 /// Cached state for a collapsed tool result, so Ctrl+O can re-render
 /// it as a fresh chamber with the full body. We hold only the last
 /// one — older collapses live in chat history but aren't addressable.
-#[derive(Clone)]
-#[allow(dead_code)]
-pub(crate) struct CollapsedToolResult {
-    pub tool_name: String,
-    pub banner_value: String,
-    pub full_output: String,
-}
-
-fn format_tool_banner_value(name: &str, args: &serde_json::Value) -> String {
-    let obj = match args {
-        serde_json::Value::Object(map) => map,
-        _ => return String::new(),
-    };
-    // `apply_patch` is structurally different: its arg is
-    // `operations: Vec<PatchOp>` (an array of ops, each with its
-    // own path), not a single string. Render "N ops" so the
-    // banner has content — degrading to bare "APPLY_PATCH" with
-    // dashes was uninformative.
-    if name == "apply_patch" {
-        let n = obj
-            .get("operations")
-            .and_then(|v| v.as_array())
-            .map(|a| a.len())
-            .unwrap_or(0);
-        return match n {
-            0 => String::new(),
-            1 => "1 op".to_string(),
-            _ => format!("{n} ops"),
-        };
-    }
-    let key = match name {
-        "read" | "write" | "edit" | "list_dir" => "path",
-        "grep" => "pattern",
-        "find_files" | "glob" => "pattern",
-        "bash" => "command",
-        "question" => "questions",
-        "task" => "prompt",
-        "task_status" => "task_id",
-        _ => return String::new(),
-    };
-    obj.get(key)
-        .and_then(|v| v.as_str())
-        .map(|s| s.trim().to_string())
-        .unwrap_or_default()
-}
-
-/// Build the rounded-chamber top border, left-truncating `value`
-/// to fill the available width up to `─╮`. Layout:
-///
-///   `╭─ TOOL ─ "value"─…─╮`
-///
-/// - When `value` fits, pad with extra `─` between the closing
-///   quote and `─╮` so the border is flush right.
-/// - When too long, take the LAST `N` chars and prefix with `…`
-///   (so filenames stay readable: paths put the filename on the
-///   right). Original PR used right-truncation, which was the
-///   wrong direction for paths.
-/// - The suffix is a tight `─╮` (no leading space) to match the
-///   chamber bottom's `╰────╯` solid-dash style. A previous
-///   version used ` ─╮` (leading space) which produced a visible
-///   gap `── ─╮` at the right edge that looked like a defect.
-fn fit_banner_header(name_upper: &str, value: &str, frame_w: usize) -> String {
-    use unicode_width::UnicodeWidthChar;
-    use unicode_width::UnicodeWidthStr;
-
-    // The header MUST be one logical line. Multi-line bash commands
-    // (line-continuation backslash + newline), tool args with
-    // embedded `\t`, or any control char that `sanitize_output`
-    // preserves would otherwise split the chamber TOP across two
-    // terminal rows — the user reported seeing `╭─ BASH ─ "...\` on
-    // row 1 and the continuation `  nikon_he..."─╮` on row 2.
-    // Collapse all whitespace runs to a single space here so the
-    // banner shape is invariant under the shape of the value.
-    let value_owned: String;
-    let value: &str = if value.contains(|c: char| c == '\n' || c == '\r' || c == '\t') {
-        value_owned = value
-            .chars()
-            .map(|c| {
-                if c == '\n' || c == '\r' || c == '\t' {
-                    ' '
-                } else {
-                    c
-                }
-            })
-            .collect::<String>()
-            // Collapse runs of spaces from the normalization above so
-            // the visible value reads as one tight line rather than
-            // `cmd \    nikon_he…` with a gap from the backslash.
-            .split_whitespace()
-            .collect::<Vec<_>>()
-            .join(" ");
-        value_owned.as_str()
-    } else {
-        value
-    };
-
-    // Cap the tool name so a pathological long name
-    // (e.g. `mcp_tool:long_server:long_function`) doesn't push
-    // the prefix past `frame_w` and overflow the chamber. Reserve
-    // enough room for the surrounding `╭─ ` + ` ─ ` + suffix +
-    // 2 quote chars + at least 1 cell of value. If the name itself
-    // is too long, left-truncate with `…` so the part closest to
-    // the colon-separated suffix (typically the function name)
-    // survives — same rationale as path truncation.
-    const FRAME_OVERHEAD: usize = 8; // "╭─ " (3) + " ─ " (3) + "─╮" (2)
-    let name_budget = frame_w.saturating_sub(FRAME_OVERHEAD + 3); // 3 = quotes + 1 value cell
-    let name_w = name_upper.width();
-    let displayed_name: String = if name_w <= name_budget || name_budget == 0 {
-        name_upper.to_string()
-    } else {
-        let tail_budget = name_budget.saturating_sub(1); // for `…`
-        let mut tail: Vec<char> = Vec::new();
-        let mut used = 0;
-        for ch in name_upper.chars().rev() {
-            let w = ch.width().unwrap_or(0);
-            if used + w > tail_budget {
-                break;
-            }
-            tail.push(ch);
-            used += w;
-        }
-        tail.reverse();
-        format!("…{}", tail.into_iter().collect::<String>())
-    };
-
-    let prefix = format!("╭─ {} ─ ", displayed_name);
-    let suffix = "─╮";
-    let prefix_w = prefix.as_str().width();
-    let suffix_w = suffix.width();
-    // Reserve at least 1 cell for padding-or-truncation marker.
-    // If frame_w is so small that even the prefix+suffix don't
-    // fit, just emit the prefix + truncated tool name with no
-    // value (degraded but doesn't panic).
-    if value.is_empty() {
-        let used = prefix_w + suffix_w;
-        let pad = frame_w.saturating_sub(used);
-        return format!("{}{}{}", prefix, "─".repeat(pad), suffix);
-    }
-    // Budget for `"value"` (the value + 2 quote chars). The
-    // chamber needs room for at least the closing suffix; if
-    // even that doesn't fit, fall back to no-value.
-    let quote_w = 2;
-    let value_budget = frame_w.saturating_sub(prefix_w + suffix_w + quote_w);
-    if value_budget == 0 {
-        let used = prefix_w + suffix_w;
-        let pad = frame_w.saturating_sub(used);
-        return format!("{}{}{}", prefix, "─".repeat(pad), suffix);
-    }
-
-    let value_w = value.width();
-    let shown_value = if value_w <= value_budget {
-        value.to_string()
-    } else {
-        // Left-truncate: take the LAST chars that fit, prefixed
-        // with `…` (1 cell). Count by display width so emoji /
-        // CJK don't break the budget.
-        use unicode_width::UnicodeWidthChar;
-        let tail_budget = value_budget.saturating_sub(1); // 1 for `…`
-        // Walk the value from the END, accumulating chars until
-        // we've used `tail_budget` cells.
-        let mut tail: Vec<char> = Vec::new();
-        let mut used = 0;
-        for ch in value.chars().rev() {
-            let w = ch.width().unwrap_or(0);
-            if used + w > tail_budget {
-                break;
-            }
-            tail.push(ch);
-            used += w;
-        }
-        tail.reverse();
-        let tail_str: String = tail.into_iter().collect();
-        format!("…{}", tail_str)
-    };
-
-    let shown_w = shown_value.as_str().width() + quote_w;
-    let total_used = prefix_w + shown_w + suffix_w;
-    let pad = frame_w.saturating_sub(total_used);
-    format!("{}\"{}\"{}{}", prefix, shown_value, "─".repeat(pad), suffix)
-}
-
 pub async fn run_interactive(
     client: AnyClient,
     mut agent: AnyAgent,
@@ -1006,7 +830,6 @@ pub async fn run_interactive(
     rewind_picker.set_monochrome(cli.no_color);
     let mut last_esc: Option<std::time::Instant> = None;
     let mut search_active = false;
-
 
     // Snapshot plugin-registered shortcuts (P9c). Seeded at UI
     // startup; refreshed at the top of each event loop iteration
@@ -4837,471 +4660,20 @@ pub async fn run_interactive(
     Ok(())
 }
 
-const ALLOW_PLACEHOLDER: &str = "<edit this pattern>";
-
 /// Whether a pattern was returned by `suggest_pattern` as the
 /// "empty input — please type a real pattern" placeholder rather
 /// than a real glob. Used by the ask-dialog to detect when the
 /// user pressed "allow always" on a degenerate input and refuse
 /// to store the placeholder as an actual allowlist entry.
 fn is_placeholder_pattern(p: &str) -> bool {
-    p == ALLOW_PLACEHOLDER
+    permission_ui::is_placeholder_pattern(p)
 }
 
 fn suggest_pattern(tool: &str, input: &str) -> String {
-    // Refuse to suggest a catch-all wildcard for empty / whitespace-
-    // only input. A user mis-clicking "(a) allow always" on an empty
-    // invocation would otherwise pin an "allow everything for this
-    // tool forever" rule into their session. The placeholder string
-    // is intentionally not a valid glob — the UI shows it as the
-    // suggested pattern, the user edits it before confirming.
-    const PLACEHOLDER: &str = ALLOW_PLACEHOLDER;
-    let trimmed = input.trim();
-    if trimmed.is_empty() {
-        return PLACEHOLDER.to_string();
-    }
-    match tool {
-        "bash" => {
-            let first = trimmed.split_whitespace().next().unwrap_or(PLACEHOLDER);
-            format!("{} *", first)
-        }
-        "read" | "write" | "edit" | "list_dir" => {
-            // Generalize to the parent dir's RECURSIVE subtree. The
-            // previous `parent/*` matched only the immediate
-            // directory (single `*` doesn't span `/` — see
-            // `permission/pattern.rs:92-96`), so after allow-always
-            // on `src/main.rs` the next write to `src/agent/foo.rs`
-            // re-prompted. Maki's `generalize_scope`
-            // (`maki-agent/src/permissions.rs:519`) uses `parent/**`
-            // for the same tools — that's what users expect when
-            // they pick "allow always".
-            let path = std::path::Path::new(trimmed);
-            let parent = path
-                .parent()
-                .map(|p| p.to_string_lossy())
-                .unwrap_or(std::borrow::Cow::Borrowed(""));
-            if parent.is_empty() {
-                "**".to_string()
-            } else {
-                format!("{}/**", parent)
-            }
-        }
-        "grep" | "find_files" => {
-            let first = trimmed.split_whitespace().next().unwrap_or(PLACEHOLDER);
-            format!("{}*", first)
-        }
-        // MCP tools: input has the shape `mcp_tool:<server>:<name>`.
-        // "Allow always" usually means "trust everything from this
-        // server" rather than "trust this exact call", so derive a
-        // wildcarded per-server pattern.
-        //
-        // Review #6: umbrella check is case-insensitive so a
-        // future caller surfacing `MCP_TOOL:...` doesn't silently
-        // hit the placeholder branch.
-        //
-        // Review #5: use rsplitn so an MCP tool with embedded
-        // colons in its name (e.g. `mcp_tool:server:do:thing`)
-        // still wildcards on the SERVER, not the whole tail. The
-        // previous `splitn(3, ':')` collapsed `do:thing` into the
-        // 3rd part — fine since we wildcard it anyway — but the
-        // pattern would over-coarsen for users wanting tool-level
-        // grants. Same behavior for the server-wildcard case
-        // either way; rsplitn just keeps the semantics
-        // deterministic if we ever switch to a different
-        // wildcarding policy.
-        "mcp_tool" => {
-            // Lowercase umbrella check; preserve original case in
-            // the emitted pattern so the allowlist matches the
-            // wire format.
-            let mut parts = trimmed.splitn(3, ':');
-            let umbrella = parts.next().unwrap_or("");
-            let server = parts.next().unwrap_or("");
-            if umbrella.eq_ignore_ascii_case("mcp_tool") && !server.is_empty() {
-                format!("mcp_tool:{}:*", server)
-            } else {
-                PLACEHOLDER.to_string()
-            }
-        }
-        // Other path-shaped tools: derive parent-dir wildcards.
-        "apply_patch" => {
-            let path = std::path::Path::new(trimmed);
-            let parent = path
-                .parent()
-                .map(|p| p.to_string_lossy())
-                .unwrap_or(std::borrow::Cow::Borrowed(""));
-            if parent.is_empty() {
-                "**".to_string()
-            } else {
-                format!("{}/**", parent)
-            }
-        }
-        // Semantic tools — file-path-shaped arguments.
-        "list_symbols" | "get_symbol_body" | "find_definition" | "find_callers"
-        | "find_callees" => {
-            let path = std::path::Path::new(trimmed);
-            let parent = path
-                .parent()
-                .map(|p| p.to_string_lossy())
-                .unwrap_or(std::borrow::Cow::Borrowed("."));
-            format!("{}/**", parent)
-        }
-        // Network tools — broad wildcard.
-        "webfetch" => "webfetch:*".to_string(),
-        "websearch" => "websearch:*".to_string(),
-        // Task / question / introspection tools — allow them entirely.
-        "task" | "task_status" | "question" => "**".to_string(),
-        "glob" | "repo_overview" | "skill" | "memory" | "write_todo_list" | "lsp" => {
-            "**".to_string()
-        }
-        _ => PLACEHOLDER.to_string(),
-    }
+    permission_ui::suggest_pattern(tool, input)
 }
 
-/// Write a line of text that must NOT land inside an open tool
-/// chamber. Closes the chamber first if any signal indicates one
-/// is open. This is the ONE entry point every tokio::select! arm
-/// should use when its output belongs "around" tool execution
-/// rather than "inside" it: lifecycle trailers, notifications,
-/// permission alerts, errors, interjection markers, "allowed …"
-/// confirmations, agent text that's not part of a tool result.
-///
-/// Structural intent — the X-inside-chamber bug has shipped at
-/// least four times (alert-in-chamber, allowed-line-in-chamber,
-/// banner-in-chamber, lifecycle-in-chamber). Each fix added
-/// `close_tool_chamber_if_open(…)` + `renderer.write_line(…)`
-/// at one specific site. This helper bundles both into one call
-/// so a NEW arm that wants to print text outside the chamber
-/// can't forget the close — using `renderer.write_line` directly
-/// inside an arm is now the suspicious choice, not the default.
-///
-/// Chamber-row paint paths (`render_tool_output`, alert frame
-/// rows, chamber TOP banner, panel sections) deliberately paint
-/// INSIDE a chamber/box and continue to call
-/// `renderer.write_line` directly. The distinction is whether
-/// the caller WANTS the chamber to be closed first.
-pub(crate) fn write_outside_chamber(
-    renderer: &mut Renderer,
-    last_tool_name: &mut Option<String>,
-    tool_chamber_open: &mut bool,
-    chamber_top_start: &mut Option<usize>,
-    chamber_top_end: &mut Option<usize>,
-    text: &str,
-    color: Color,
-) -> anyhow::Result<()> {
-    // Passive close — just terminate the visual frame. The tool
-    // isn't being denied; we're moving on. Active denial paths
-    // (permission deny, agent error, interjection) call
-    // `close_tool_chamber_abort` themselves so they get the
-    // "⚠ tool denied" wording.
-    //
-    // Defense-in-depth (review #7): sanitize control bytes here
-    // so a future caller that forgets to sanitize can't smuggle
-    // ANSI escapes into chat. KEEP_NEWLINE preserves intentional
-    // multi-line content; `renderer.write_line` handles the
-    // per-line splits + per-row caps.
-    close_tool_chamber_passive(
-        renderer,
-        last_tool_name,
-        tool_chamber_open,
-        chamber_top_start,
-        chamber_top_end,
-    )?;
-    let safe = crate::ui::ansi::strip_controls(text, crate::ui::ansi::StripPolicy::KEEP_NEWLINE);
-    renderer.write_line(&safe, color)?;
-    Ok(())
-}
-
-/// Close an in-flight chamber WITH an abort/denied row painted
-/// inside. Use this when the chamber is closing because the tool
-/// was actively rejected — permission deny, agent error,
-/// interjection. The user needs to see WHY no result is coming.
-fn close_tool_chamber_abort(
-    renderer: &mut Renderer,
-    last_tool_name: &mut Option<String>,
-    tool_chamber_open: &mut bool,
-) -> anyhow::Result<()> {
-    if last_tool_name.is_some() || *tool_chamber_open {
-        let (frame_w, inner) = chamber_widths(renderer);
-        renderer.write_line(
-            &chamber_row_centered("⚠ tool denied · aborted · no result", inner),
-            theme::perm(),
-        )?;
-        renderer.write_line(&chamber_bottom(frame_w), theme::dim())?;
-        *last_tool_name = None;
-        *tool_chamber_open = false;
-    }
-    Ok(())
-}
-
-/// Close an in-flight chamber WITHOUT painting an abort row —
-/// just emit the chamber bottom and clear state. Use this when
-/// the chamber is being closed because something ELSE (notification,
-/// lifecycle event, question prompt, plan switch, plugin dialog)
-/// wants to print outside the chamber. The tool itself isn't
-/// denied; we just need to terminate the visual frame so the new
-/// content doesn't land inside.
-///
-/// Reviewer-caught regression: the previous "structural fix"
-/// reused `close_tool_chamber_if_open` (always painted "⚠ tool
-/// denied"). Wiring that under `write_outside_chamber` meant every
-/// notification arriving mid-chamber would falsely brand the tool
-/// as denied. The split lets each caller pick the right wording.
-fn close_tool_chamber_passive(
-    renderer: &mut Renderer,
-    last_tool_name: &mut Option<String>,
-    tool_chamber_open: &mut bool,
-    chamber_top_start: &mut Option<usize>,
-    chamber_top_end: &mut Option<usize>,
-) -> anyhow::Result<()> {
-    if last_tool_name.is_some() || *tool_chamber_open {
-        // If the chamber TOP was painted but no body content
-        // followed (parallel tool-call displacement: the agent
-        // fired the next tool before any result for this one
-        // landed), DROP the chamber entirely instead of closing
-        // an empty box. The result for the displaced id will
-        // paint its own fresh chamber via the dirge-jzj path.
-        let drop_chamber = match (*chamber_top_start, *chamber_top_end) {
-            (Some(_start), Some(end)) => renderer.buffer_len() == end,
-            _ => false,
-        };
-        if drop_chamber {
-            if let Some(start) = *chamber_top_start {
-                renderer.replace_from(start, Vec::new());
-            }
-        } else {
-            let (frame_w, _inner) = chamber_widths(renderer);
-            renderer.write_line(&chamber_bottom(frame_w), theme::dim())?;
-        }
-        *last_tool_name = None;
-        *tool_chamber_open = false;
-        *chamber_top_start = None;
-        *chamber_top_end = None;
-    }
-    Ok(())
-}
-
-/// Back-compat alias for the abort variant. Every existing
-/// caller of `close_tool_chamber_if_open` was in an
-/// abort-shaped context (permission deny / error / interjection /
-/// fresh tool over stale chamber), so the aliasing preserves
-/// their behavior. New code should pick the variant explicitly.
-fn close_tool_chamber_if_open(
-    renderer: &mut Renderer,
-    last_tool_name: &mut Option<String>,
-    tool_chamber_open: &mut bool,
-) -> anyhow::Result<()> {
-    close_tool_chamber_abort(renderer, last_tool_name, tool_chamber_open)
-}
-
-/// `│   <content centered to inner>   │` — pad text on both sides so
-/// it sits horizontally centered within the chamber inner width.
-fn chamber_row_centered(content: &str, inner: usize) -> String {
-    // Total row width matches `chamber_row`: exactly `inner + 4`
-    // terminal cells (`│ ` (2) + inner-cell middle + ` │` (2)).
-    // The middle is `inner` cells: left_pad + content + right_pad.
-    //
-    // TWO bugs were stacked here:
-    // (1) Used `chars().count()` instead of display width — the
-    //     NO-OUTPUT chamber starts with `⚠` (2 cells / 1 char) so
-    //     centering was off by 1 cell.
-    // (2) Subtracted `len + 2` from `inner` for the pad, leaving
-    //     the row 2 cells short of `inner + 4` total — the right
-    //     `│` didn't line up under the chamber's top `╮` /
-    //     bottom `╯`. Correct: pad = inner - len (the middle is
-    //     `inner` cells wide; subtracting only `len` reserves the
-    //     rest for padding around it).
-    //
-    // Anything wider than `inner` falls back to `chamber_row`
-    // which truncates with `…` and pads to exactly `inner` cells.
-    use unicode_width::UnicodeWidthStr;
-    let len = UnicodeWidthStr::width(content);
-    if len >= inner {
-        return chamber_row(content, inner);
-    }
-    let pad = inner - len;
-    let left = pad / 2;
-    let right = pad - left;
-    format!("│ {}{}{} │", " ".repeat(left), content, " ".repeat(right))
-}
-
-/// Tools whose result body is the WHOLE point. Collapsing these to
-/// 4 lines hides the value the user just asked for; keep them
-/// fully expanded.
-///
-/// - `edit` — colorized diff with backgrounds; the diff IS the value.
-/// - `question` — Q&A response; collapsing hides Anwers past Q1.
-/// - `task` / `task_status` — subagent results.
-/// - `read` — when the LLM (or user via /read) asked to read a file,
-///   they want to see the file content; defaulting to 4 lines defeats
-///   the request.
-///
-/// `apply_patch` deliberately NOT exempt: its output is usually a
-/// short "N ops applied" summary that fits in 4 lines; on the rare
-/// failure with per-op errors, the collapse + Ctrl+O is the right
-/// affordance. Reviewer feedback led to this trim.
-fn tool_skips_collapse(tool_name: &str) -> bool {
-    matches!(
-        tool_name,
-        "edit" | "read" | "question" | "task" | "task_status"
-    )
-}
-
-/// Render a tool result chamber. Lines past `max_lines` collapse to a
-/// `↓ N more lines (Ctrl+O to expand)` footer; `max_chars` is a hard
-/// ceiling on the displayed slice. `max_lines = usize::MAX` (passed
-/// for tools in `tool_skips_collapse`) disables the line cap.
-/// Returns `Some(CollapsedToolResult)` if the body was actually
-/// truncated, so the caller can stash it for Ctrl+O.
-fn render_tool_output(
-    renderer: &mut Renderer,
-    tool_name: &str,
-    banner_value: &str,
-    output: &str,
-    max_chars: usize,
-    max_lines: usize,
-) -> anyhow::Result<Option<CollapsedToolResult>> {
-    let sanitized = sanitize_output(output);
-    let total_chars = sanitized.chars().count();
-    // Hard char ceiling first — keeps a single pathological line
-    // from blowing the chamber even if the line count is fine.
-    let char_sliced: String = if total_chars <= max_chars {
-        sanitized.into_string()
-    } else {
-        sanitized.chars().take(max_chars).collect()
-    };
-    let chars_truncated = total_chars.saturating_sub(char_sliced.chars().count());
-
-    // Apply the per-tool line cap (skipped for diff / Q&A / task
-    // results where the body IS the value).
-    let lines: Vec<&str> = char_sliced.lines().collect();
-    let total_lines = lines.len();
-    let line_cap = if tool_skips_collapse(tool_name) {
-        usize::MAX
-    } else {
-        max_lines
-    };
-    let shown_lines = total_lines.min(line_cap);
-    let hidden_lines = total_lines.saturating_sub(shown_lines);
-
-    let (frame_w, inner) = chamber_widths(renderer);
-    // When the tool returned nothing (or only whitespace), paint a
-    // dim placeholder row so the chamber never appears as a bare
-    // ╭───╮ + ╰───╯ pair. Per-tool wording so it reads naturally
-    // — `glob`/`grep` say "no matches", `read` says "empty file",
-    // everything else falls back to "(no output)".
-    let body_is_empty = char_sliced.trim().is_empty();
-    if body_is_empty {
-        let placeholder = match tool_name {
-            "glob" | "grep" | "find" | "semantic" => "(no matches)",
-            "read" => "(empty file)",
-            "bash" => "(no output)",
-            _ => "(no output)",
-        };
-        renderer.write_line(&chamber_row(placeholder, inner), theme::dim())?;
-    }
-    for line in &lines[..shown_lines] {
-        renderer.write_line(&chamber_row(line, inner), theme::result())?;
-    }
-    if hidden_lines > 0 {
-        // The Ctrl+O hint inside the row — concise enough to fit the
-        // 120-col chamber without truncation on typical widths.
-        let note = format!(
-            "↓ {} more line{} (Ctrl+O to expand)",
-            hidden_lines,
-            if hidden_lines == 1 { "" } else { "s" }
-        );
-        renderer.write_line(&chamber_row(&note, inner), theme::dim())?;
-    }
-    if chars_truncated > 0 {
-        // Char-ceiling hit BEFORE the line cap. Report separately so
-        // the user knows the char-truncated text has already been
-        // dropped (expanding via Ctrl+O won't bring it back).
-        let note = format!("░ +{} chars truncated (output too large)", chars_truncated);
-        renderer.write_line(&chamber_row(&note, inner), theme::dim())?;
-    }
-    renderer.write_line(&chamber_bottom(frame_w), theme::dim())?;
-
-    // Review #3 + #8: stash for Ctrl+O when EITHER line OR char
-    // truncation hid data; the char-only case previously returned
-    // `None` and Ctrl+O reported "nothing to expand" despite the
-    // visible `+N chars truncated` footer. Sanitize `banner_value`
-    // here once — the expand path re-emits the banner through
-    // `fit_banner_header` and would otherwise paint raw ANSI from
-    // MCP / attacker-shaped args at expand time.
-    if hidden_lines > 0 || chars_truncated > 0 {
-        Ok(Some(CollapsedToolResult {
-            tool_name: tool_name.to_string(),
-            banner_value: sanitize_output(banner_value).into_string(),
-            full_output: output.to_string(),
-        }))
-    } else {
-        Ok(None)
-    }
-}
-
-/// Re-render a previously-collapsed result with NO line cap, as a
-/// fresh chamber below the prior chat content. Char ceiling still
-/// applies so a multi-MB output doesn't take forever.
-#[allow(dead_code)]
-fn render_collapsed_in_full(
-    renderer: &mut Renderer,
-    collapsed: &CollapsedToolResult,
-    max_chars: usize,
-) -> anyhow::Result<()> {
-    let upper = collapsed.tool_name.to_ascii_uppercase();
-    let (frame_w, _) = chamber_widths(renderer);
-    let header = fit_banner_header(&upper, &collapsed.banner_value, frame_w);
-    renderer.write_line("", Color::White)?;
-    renderer.write_line(&header, theme::tool())?;
-    // Skip-collapse semantics here so even a normally-collapsed tool
-    // (e.g. read) reprints its body in full when the user opts in.
-    let _ = render_tool_output(
-        renderer,
-        &collapsed.tool_name,
-        &collapsed.banner_value,
-        &collapsed.full_output,
-        max_chars,
-        usize::MAX,
-    )?;
-    Ok(())
-}
-
-/// Standard tool-chamber widths derived from the renderer's content
-/// area. Capped at 120 so very wide terminals don't produce sprawling
-/// chambers that overwhelm the content.
-fn chamber_widths(renderer: &Renderer) -> (usize, usize) {
-    // Match ChatPane's 1-cell right margin EXACTLY. ChatPane paints
-    // chat content into `chat.width - 1` cells (the last cell is a
-    // reserved margin before the ║ border). Chamber lines pass
-    // through that same painter, so any chamber wider than
-    // `content_width - 1` gets its right `╮`/`╯` corner clipped.
-    let frame_w = renderer.content_width().saturating_sub(1).max(20);
-    let inner = frame_w.saturating_sub(4); // `│ ` + ` │`
-    (frame_w, inner)
-}
-
-/// `╰────────────╯` footer of a tool chamber. Thin wrapper around
-/// `box_render::bottom` so chamber + alert + panel sections all
-/// share one frame implementation.
-fn chamber_bottom(frame_w: usize) -> String {
-    box_render::bottom(box_render::BoxStyle::Rounded, frame_w)
-}
-
-/// `│ content │` row of a tool chamber. `inner` is the
-/// content cells (frame_w − 4 for the two borders + spaces). Thin
-/// wrapper around `box_render::row` which does the same display-
-/// width / tab-expansion / truncate-with-`…` logic.
-fn chamber_row(content: &str, inner: usize) -> String {
-    box_render::row(box_render::BoxStyle::Rounded, content, inner + 4)
-}
-
-/// Background-tinted chamber row for diff `+`/`-` lines. Thin
-/// wrapper around `box_render::row_with_bg`.
-fn chamber_row_with_bg(content: &str, inner: usize, bg_idx: u8) -> String {
-    box_render::row_with_bg(box_render::BoxStyle::Rounded, content, inner + 4, bg_idx)
-}
-
-/// Ctrl-F scrollback search. Ported from maki
+/// Whether a pattern was returned by `suggest_pattern` as the
 /// (`maki-ui/src/components/search_modal.rs:147-185`): fuzzy match via
 /// `nucleo-matcher`, ranked by score descending so the best matches
 /// surface first. Previously this was a `to_lowercase().contains()`
@@ -5750,609 +5122,7 @@ mod tests {
         }));
     }
 
-    /// Chamber banner fills the full frame width when the value fits.
-    /// Right border is flush at frame_w; no extra trailing whitespace.
-    #[test]
-    fn banner_short_value_pads_with_dashes_to_full_width() {
-        let header = fit_banner_header("READ", "/tmp/x", 60);
-        assert_eq!(
-            header.as_str().width(),
-            60,
-            "header should fill frame_w exactly: {:?}",
-            header,
-        );
-        assert!(header.starts_with("╭─ READ ─ \"/tmp/x\""));
-        assert!(header.ends_with("─╮"));
-    }
-
-    /// Regression: the suffix must be a tight `─╮` (no leading
-    /// space). The previous ` ─╮` form left a visible gap
-    /// `── ─╮` at the right edge that looked like the dash run
-    /// was broken. Solid dashes match the chamber bottom
-    /// (`╰────╯`) style.
-    #[test]
-    fn banner_has_no_internal_space_before_corner() {
-        let header = fit_banner_header("READ", "/short", 50);
-        // Strip the closing ╮ and check the char just before it
-        // is `─`, not ` `.
-        let mut chars: Vec<char> = header.chars().collect();
-        let last = chars.pop();
-        assert_eq!(last, Some('╮'));
-        let second_last = chars.pop();
-        assert_eq!(
-            second_last,
-            Some('─'),
-            "char before closing ╮ must be `─`, not space; got {:?}",
-            second_last,
-        );
-    }
-
-    /// Banner left-truncates long paths so the filename (right side
-    /// of a path) stays visible. The prefix `…` signals truncation.
-    #[test]
-    fn banner_long_path_left_truncates_to_preserve_filename() {
-        let path = "/very/very/very/long/nested/path/to/some/file/named/important.clj";
-        let header = fit_banner_header("READ", path, 60);
-        assert_eq!(
-            header.as_str().width(),
-            60,
-            "header must still be exactly frame_w wide",
-        );
-        // Filename must be visible (right side of path was preserved).
-        assert!(
-            header.contains("important.clj"),
-            "filename should survive truncation: {:?}",
-            header,
-        );
-        // Truncation marker is present.
-        assert!(header.contains('…'), "expected `…` marker: {:?}", header);
-        // The `/very/very` head SHOULD be gone — that's the whole
-        // point of left-truncation.
-        assert!(
-            !header.contains("/very/very/very/long"),
-            "leading head should be truncated: {:?}",
-            header,
-        );
-    }
-
-    /// Regression: bash line-continuation commands carry a literal
-    /// `\n` through `sanitize_output` (which preserves `\n`/`\t` so
-    /// chat content stays whole). The banner header MUST collapse
-    /// those into a single visible line — otherwise `write_line`
-    /// splits on `\n` and emits two terminal rows, breaking the
-    /// chamber TOP into `╭─ BASH ─ "...\` on row 1 and
-    /// `  rest..."─╮` on row 2.
-    #[test]
-    fn banner_collapses_embedded_newlines_to_single_line() {
-        let multi = "clang++ predecessor.cpp \\\n  nikon_he_precinct_decode.cpp 2>&1";
-        let header = fit_banner_header("BASH", multi, 80);
-        assert!(
-            !header.contains('\n'),
-            "header must not contain newlines: {:?}",
-            header,
-        );
-        assert!(
-            !header.contains('\t'),
-            "header must not contain tabs: {:?}",
-            header,
-        );
-        assert!(
-            !header.contains('\r'),
-            "header must not contain carriage returns: {:?}",
-            header,
-        );
-        // Width invariant still holds.
-        assert_eq!(header.as_str().width(), 80);
-    }
-
-    /// Banner with a tab in the value (Makefile targets, etc.) also
-    /// collapses to a single line.
-    #[test]
-    fn banner_collapses_embedded_tabs() {
-        let header = fit_banner_header("READ", "path\twith\ttabs", 60);
-        assert!(!header.contains('\t'));
-        assert_eq!(header.as_str().width(), 60);
-    }
-
-    /// Empty value (e.g. tool with no banner-worthy arg) renders
-    /// just the prefix with dash fill — no empty quote pair.
-    #[test]
-    fn banner_empty_value_renders_just_prefix_and_dashes() {
-        let header = fit_banner_header("DONE", "", 50);
-        assert_eq!(header.as_str().width(), 50);
-        assert!(
-            !header.contains("\"\""),
-            "no empty quote pair: {:?}",
-            header
-        );
-        assert!(header.starts_with("╭─ DONE ─"));
-        assert!(header.ends_with("─╮"));
-    }
-
-    /// Regression: `chamber_row_centered` must use DISPLAY width
-    /// not char count. The NO-OUTPUT chamber message starts with
-    /// `⚠` (2 cells wide, 1 char). Before this, centering was off
-    /// by 1 cell and the right `│` border misaligned with the
-    /// chamber's top/bottom borders.
-    #[test]
-    fn chamber_row_centered_handles_wide_emoji() {
-        let row = chamber_row_centered("⚠ tool denied", 40);
-        // Row must occupy exactly `inner + 4` display cells
-        // (`│ ` + content + padding + ` │` = inner + 4 = 44).
-        let row_width = UnicodeWidthStr::width(row.as_str());
-        assert_eq!(
-            row_width, 44,
-            "row must be exactly inner+4 cells wide; got {row_width} for {row:?}",
-        );
-        // Right border `│` MUST be at the very end (no trailing
-        // pad mismatch).
-        assert!(
-            row.ends_with(" │"),
-            "right border missing or padded wrong: {row:?}"
-        );
-    }
-
-    /// `chamber_row` must align by display width too, not `char` count.
-    /// Tool output rows (`render_tool_output`) flow through this, so
-    /// any CJK/emoji in a `read`/`bash`/`grep` result would push the
-    /// right `│` past the chamber's column. Sibling
-    /// `chamber_row_centered` was already fixed; this is the same
-    /// bug in the asymmetric helper.
-    #[test]
-    fn chamber_row_handles_wide_emoji() {
-        // Short content with one wide char: must pad to exactly
-        // inner+4 cells. "ok ✅" = 2+1+2 = 5 cells (4 chars).
-        let row = chamber_row("ok ✅", 40);
-        let row_width = UnicodeWidthStr::width(row.as_str());
-        assert_eq!(
-            row_width, 44,
-            "row must be exactly inner+4 cells wide; got {row_width} for {row:?}",
-        );
-        assert!(row.ends_with(" │"), "right border missing: {row:?}");
-
-        // Long content with mixed wide chars: must truncate by
-        // display width and still land at exactly inner+4 cells.
-        let long = "日本語日本語日本語日本語日本語日本語日本語日本語日本語日本語";
-        let row = chamber_row(long, 20);
-        let row_width = UnicodeWidthStr::width(row.as_str());
-        assert_eq!(
-            row_width, 24,
-            "wide-char row must be exactly inner+4 cells wide; got {row_width} for {row:?}",
-        );
-        assert!(
-            row.ends_with(" │"),
-            "right border missing on truncated wide-char row: {row:?}"
-        );
-    }
-
-    /// Regression for the "ALERT renders inside open chamber" report:
-    /// `close_tool_chamber_if_open` must close on EITHER signal —
-    /// the legacy `last_tool_name.is_some()` OR the newer
-    /// `tool_chamber_open` flag. The two state variables can
-    /// disagree because `last_tool_name` is cleared by paths that
-    /// don't paint a chamber bottom (Done, etc.), so a chamber TOP
-    /// can be on-screen while the name slot is None. Without the OR
-    /// gate the alert handler would skip the close and the alert
-    /// would render directly under an unclosed `╭─ TOOL ─` TOP.
-    #[test]
-    fn close_tool_chamber_fires_when_only_flag_is_open() {
-        let mut renderer = Renderer::new().expect("renderer");
-        // Pretend a chamber was painted but `last_tool_name` was
-        // drained by an unrelated clear site — exactly the bug
-        // shape from the user's screenshot.
-        let mut name: Option<String> = None;
-        let mut open = true;
-        close_tool_chamber_if_open(&mut renderer, &mut name, &mut open).unwrap();
-        assert!(!open, "flag must be cleared by the close");
-        assert!(name.is_none(), "name stays cleared after the close");
-        // Symmetric case: flag false, name Some → also closes
-        // (legacy behavior preserved).
-        let mut name: Option<String> = Some("read".to_string());
-        let mut open = false;
-        close_tool_chamber_if_open(&mut renderer, &mut name, &mut open).unwrap();
-        assert!(name.is_none(), "name cleared on legacy-signal close");
-        assert!(!open, "flag stays cleared");
-        // Both false → no-op (idempotent).
-        let mut name: Option<String> = None;
-        let mut open = false;
-        close_tool_chamber_if_open(&mut renderer, &mut name, &mut open).unwrap();
-        assert!(name.is_none());
-        assert!(!open);
-    }
-
-    /// Structural test: `write_outside_chamber` closes any open
-    /// chamber FIRST, then writes the requested line. The chamber
-    /// state flags are cleared regardless of which signal was
-    /// raised — same semantics as a manual `close_tool_chamber_if_open`
-    /// followed by `renderer.write_line`, just bundled so future
-    /// tokio::select! arms can't ship the X-inside-chamber bug by
-    /// forgetting the close call.
-    #[test]
-    fn write_outside_chamber_closes_chamber_first() {
-        let mut renderer = Renderer::new().expect("renderer");
-        // Flag-only open (the regression case from past bug
-        // reports).
-        let mut name: Option<String> = None;
-        let mut open = true;
-        let mut start: Option<usize> = None;
-        let mut end: Option<usize> = None;
-        write_outside_chamber(
-            &mut renderer,
-            &mut name,
-            &mut open,
-            &mut start,
-            &mut end,
-            "hello",
-            Color::White,
-        )
-        .unwrap();
-        assert!(!open, "chamber must be closed");
-        assert!(name.is_none());
-
-        // Name-only open (legacy signal).
-        let mut name: Option<String> = Some("read".to_string());
-        let mut open = false;
-        let mut start: Option<usize> = None;
-        let mut end: Option<usize> = None;
-        write_outside_chamber(
-            &mut renderer,
-            &mut name,
-            &mut open,
-            &mut start,
-            &mut end,
-            "hi",
-            Color::White,
-        )
-        .unwrap();
-        assert!(name.is_none());
-        assert!(!open);
-
-        // No chamber open — write_outside_chamber still works
-        // (idempotent close, then write).
-        let mut name: Option<String> = None;
-        let mut open = false;
-        let mut start: Option<usize> = None;
-        let mut end: Option<usize> = None;
-        write_outside_chamber(
-            &mut renderer,
-            &mut name,
-            &mut open,
-            &mut start,
-            &mut end,
-            "plain",
-            Color::White,
-        )
-        .unwrap();
-        // No assertion on state — just verifying it doesn't panic
-        // / error.
-    }
-
-    /// Reviewer-caught regression: previously `write_outside_chamber`
-    /// reused `close_tool_chamber_if_open` which ALWAYS painted
-    /// "⚠ tool denied · aborted · no result" — meaning every
-    /// notification arriving mid-chamber would falsely brand the
-    /// in-flight tool as denied. After the abort/passive split,
-    /// the helper uses the passive close which only emits the
-    /// chamber bottom + clears state.
-    #[test]
-    fn close_passive_drops_empty_chamber() {
-        // When the chamber TOP was painted but no body content
-        // followed, passive close DROPS the chamber entirely
-        // (truncates buffer back to before the TOP). No empty
-        // box on screen.
-        let mut renderer = Renderer::new().expect("renderer");
-        let chamber_start = renderer.buffer_len();
-        // Simulate a ToolCall painting spacer + header.
-        renderer.write_line("", Color::White).unwrap();
-        renderer
-            .write_line("╭─ MOCK_TOOL ─────╮", Color::White)
-            .unwrap();
-        let chamber_end = renderer.buffer_len();
-        assert_eq!(chamber_end - chamber_start, 2);
-
-        let mut name: Option<String> = Some("mock".into());
-        let mut open = true;
-        let mut start: Option<usize> = Some(chamber_start);
-        let mut end: Option<usize> = Some(chamber_end);
-        close_tool_chamber_passive(&mut renderer, &mut name, &mut open, &mut start, &mut end)
-            .unwrap();
-        // Buffer is back to BEFORE the chamber TOP — the empty
-        // chamber is gone entirely.
-        assert_eq!(
-            renderer.buffer_len(),
-            chamber_start,
-            "passive close on an empty chamber should drop it"
-        );
-        assert!(!open);
-        assert!(name.is_none());
-        assert!(start.is_none());
-        assert!(end.is_none());
-    }
-
-    /// When body content WAS added (buffer_len > chamber_top_end),
-    /// passive close paints a bottom border to close the chamber
-    /// normally — does NOT truncate.
-    #[test]
-    fn close_passive_with_body_writes_bottom() {
-        let mut renderer = Renderer::new().expect("renderer");
-        let chamber_start = renderer.buffer_len();
-        renderer.write_line("", Color::White).unwrap();
-        renderer
-            .write_line("╭─ MOCK_TOOL ─────╮", Color::White)
-            .unwrap();
-        let chamber_end = renderer.buffer_len();
-        // Add body content.
-        renderer.write_line("│ body row 1 │", Color::White).unwrap();
-        let after_body = renderer.buffer_len();
-        assert!(after_body > chamber_end);
-
-        let mut name: Option<String> = Some("mock".into());
-        let mut open = true;
-        let mut start: Option<usize> = Some(chamber_start);
-        let mut end: Option<usize> = Some(chamber_end);
-        close_tool_chamber_passive(&mut renderer, &mut name, &mut open, &mut start, &mut end)
-            .unwrap();
-        // One more row (the chamber bottom) is appended — no truncation.
-        assert_eq!(
-            renderer.buffer_len(),
-            after_body + 1,
-            "passive close with body should append exactly one row (bottom)",
-        );
-        let body_lines = renderer.buffer_lines();
-        assert!(body_lines.last().unwrap().contains('╰'));
-        assert!(!open);
-        assert!(name.is_none());
-    }
-
-    /// Abort variant still emits two rows (warning + bottom) so
-    /// permission-deny / error / interjection paths keep their
-    /// existing wording.
-    #[test]
-    fn close_abort_paints_warning_and_bottom() {
-        let mut renderer = Renderer::new().expect("renderer");
-        let initial_buffer_len = renderer.buffer_len();
-        let mut name: Option<String> = None;
-        let mut open = true;
-        close_tool_chamber_abort(&mut renderer, &mut name, &mut open).unwrap();
-        let after = renderer.buffer_len();
-        assert_eq!(
-            after - initial_buffer_len,
-            2,
-            "abort close should emit warning row + bottom border",
-        );
-        assert!(!open);
-        assert!(name.is_none());
-    }
-
-    /// Tool-result body collapse: long output truncates at
-    /// `max_lines` and returns a `CollapsedToolResult` so Ctrl+O
-    /// can replay it. Tools in `tool_skips_collapse` bypass the
-    /// line cap (returns None).
-    #[test]
-    fn render_tool_output_collapses_past_max_lines() {
-        let mut renderer = Renderer::new().expect("renderer");
-        let output = (0..20)
-            .map(|i| format!("line {i}"))
-            .collect::<Vec<_>>()
-            .join("\n");
-        // Pick a tool NOT in tool_skips_collapse — `grep` qualifies.
-        // (`read` was moved to the exempt set per reviewer feedback.)
-        let collapsed = render_tool_output(&mut renderer, "grep", "pattern", &output, 10_000, 4)
-            .expect("render ok");
-        let c = collapsed.expect("grep should collapse past 4 lines");
-        assert_eq!(c.tool_name, "grep");
-        assert_eq!(c.banner_value, "pattern");
-        assert!(c.full_output.contains("line 19"));
-    }
-
-    /// Exempt tools (edit / read / question / task / task_status)
-    /// pass the line cap through unchanged — `tool_skips_collapse`
-    /// forces `usize::MAX`. The render call returns None (no
-    /// collapse store) even when the output is 20 lines long.
-    /// `apply_patch` is deliberately NOT in the exempt set.
-    #[test]
-    fn render_tool_output_does_not_collapse_exempt_tools() {
-        let mut renderer = Renderer::new().expect("renderer");
-        let output = (0..20)
-            .map(|i| format!("+ added line {i}"))
-            .collect::<Vec<_>>()
-            .join("\n");
-        for tool in ["edit", "read", "question", "task", "task_status"] {
-            let collapsed = render_tool_output(&mut renderer, tool, "arg", &output, 10_000, 4)
-                .expect("render ok");
-            assert!(
-                collapsed.is_none(),
-                "exempt tool `{}` must not collapse",
-                tool,
-            );
-        }
-    }
-
-    /// `apply_patch` was removed from the exempt set (reviewer
-    /// feedback). A long apply_patch output should collapse + stash
-    /// so Ctrl+O works on per-op error spew.
-    #[test]
-    fn render_tool_output_apply_patch_collapses() {
-        let mut renderer = Renderer::new().expect("renderer");
-        let output = (0..20)
-            .map(|i| format!("op {i} applied"))
-            .collect::<Vec<_>>()
-            .join("\n");
-        let collapsed =
-            render_tool_output(&mut renderer, "apply_patch", "20 ops", &output, 10_000, 4)
-                .expect("render ok");
-        assert!(
-            collapsed.is_some(),
-            "apply_patch must collapse past max_lines",
-        );
-    }
-
-    /// Review #3: char-truncated output whose char-sliced form fits
-    /// inside `max_lines` still has data hidden — `chars_truncated >
-    /// 0` even though `hidden_lines == 0`. Before the fix, this
-    /// returned `None` and Ctrl+O reported "nothing to expand"
-    /// while the `+N chars truncated` footer was visibly hinting at
-    /// hidden content. The expected behavior: stash so Ctrl+O can
-    /// expand to the full output.
-    #[test]
-    fn render_tool_output_stashes_on_char_truncation_alone() {
-        let mut renderer = Renderer::new().expect("renderer");
-        // Single very long line — line cap (4) is irrelevant, only
-        // char cap fires. Pick a tool that ISN'T in
-        // tool_skips_collapse so the cap engages.
-        let long_single_line = "a".repeat(50_000);
-        let collapsed = render_tool_output(
-            &mut renderer,
-            "grep",
-            "pattern",
-            &long_single_line,
-            500, // hard char cap
-            4,
-        )
-        .expect("render ok");
-        let c = collapsed.expect("char-truncation alone must still stash for Ctrl+O");
-        assert_eq!(c.full_output.len(), 50_000);
-    }
-
-    /// Empty tool output gets a `(no matches)` / `(empty file)` /
-    /// `(no output)` placeholder row so the chamber never appears
-    /// as a bare ╭───╮ ╰───╯ with nothing between.
-    #[test]
-    fn render_tool_output_empty_body_gets_placeholder() {
-        // glob with no matches → "(no matches)" placeholder.
-        let mut renderer = Renderer::new().expect("renderer");
-        render_tool_output(&mut renderer, "glob", "**/*.nonexistent", "", 10_000, 100)
-            .expect("render ok");
-        let body_text: Vec<&str> = renderer.buffer_lines();
-        assert!(
-            body_text.iter().any(|l| l.contains("(no matches)")),
-            "glob with empty output should show '(no matches)'; got {:?}",
-            body_text
-        );
-
-        // read with whitespace-only file → "(empty file)".
-        let mut renderer = Renderer::new().expect("renderer");
-        render_tool_output(&mut renderer, "read", "empty.txt", "   \n\n  ", 10_000, 100)
-            .expect("render ok");
-        let body_text: Vec<&str> = renderer.buffer_lines();
-        assert!(
-            body_text.iter().any(|l| l.contains("(empty file)")),
-            "read with whitespace-only output should show '(empty file)'; got {:?}",
-            body_text
-        );
-
-        // Unknown tool → generic "(no output)".
-        let mut renderer = Renderer::new().expect("renderer");
-        render_tool_output(&mut renderer, "weird_tool", "x", "", 10_000, 100).expect("render ok");
-        let body_text: Vec<&str> = renderer.buffer_lines();
-        assert!(
-            body_text.iter().any(|l| l.contains("(no output)")),
-            "unknown tool with empty output should show '(no output)'; got {:?}",
-            body_text
-        );
-    }
-
-    /// Output that fits in `max_lines` returns None (no collapse
-    /// indicator) — no expand footer rendered, and no entry stashed
-    /// for Ctrl+O so the keybind correctly reports "nothing to
-    /// expand" if pressed.
-    #[test]
-    fn render_tool_output_returns_none_when_no_truncation() {
-        let mut renderer = Renderer::new().expect("renderer");
-        let collapsed = render_tool_output(
-            &mut renderer,
-            "list_dir",
-            ".",
-            "1 entries (1 files):\n  [file]  foo.txt",
-            10_000,
-            4,
-        )
-        .expect("render ok");
-        assert!(collapsed.is_none());
-    }
-
-    /// Self-review bug 1: `apply_patch` arg is `operations`
-    /// (array), not a single string. Previously fell through to
-    /// "path" lookup which returned empty, degrading the banner
-    /// to bare "APPLY_PATCH" with dashes. Now shows op count.
-    #[test]
-    fn banner_value_apply_patch_shows_op_count() {
-        let args = serde_json::json!({"operations": [{"action": "create", "path": "/a"}]});
-        assert_eq!(format_tool_banner_value("apply_patch", &args), "1 op");
-
-        let args = serde_json::json!({
-            "operations": [
-                {"action": "create", "path": "/a"},
-                {"action": "update", "path": "/b"},
-                {"action": "delete", "path": "/c"},
-            ],
-        });
-        assert_eq!(format_tool_banner_value("apply_patch", &args), "3 ops");
-
-        // Empty operations array → empty value (banner degrades
-        // gracefully to prefix + dashes + suffix).
-        let args = serde_json::json!({"operations": []});
-        assert_eq!(format_tool_banner_value("apply_patch", &args), "");
-
-        // Missing operations key → empty.
-        let args = serde_json::json!({});
-        assert_eq!(format_tool_banner_value("apply_patch", &args), "");
-    }
-
-    /// `format_tool_banner_value` picks the right key per tool.
-    #[test]
-    fn banner_value_picks_correct_key_per_tool() {
-        let args =
-            serde_json::json!({"path": "/p", "command": "ls", "pattern": "*.rs", "task_id": "t1"});
-        assert_eq!(format_tool_banner_value("read", &args), "/p");
-        assert_eq!(format_tool_banner_value("write", &args), "/p");
-        assert_eq!(format_tool_banner_value("edit", &args), "/p");
-        assert_eq!(format_tool_banner_value("bash", &args), "ls");
-        assert_eq!(format_tool_banner_value("grep", &args), "*.rs");
-        assert_eq!(format_tool_banner_value("glob", &args), "*.rs");
-        assert_eq!(format_tool_banner_value("task_status", &args), "t1");
-        // Unknown tool → empty (header degrades to prefix-only).
-        assert_eq!(format_tool_banner_value("mystery", &args), "");
-    }
-
-    /// Edge: very narrow frame_w that can't fit even one value
-    /// char. Must degrade gracefully without panicking; just emits
-    /// prefix + dashes + suffix.
-    #[test]
-    fn banner_handles_pathologically_narrow_frame() {
-        // Just enough for "╭─ X ─ " + " ─╮" + maybe nothing else.
-        let header = fit_banner_header("READ", "/some/path", 12);
-        // Don't pin the exact string — just make sure no panic and
-        // we got SOMETHING with the borders intact.
-        assert!(header.starts_with("╭"));
-        assert!(header.ends_with("╮"));
-    }
-
-    /// Regression: a very long tool name (e.g. an MCP-registered
-    /// tool like `MCP_TOOL:LONG_SERVER:LONG_FUNCTION`) used to
-    /// overflow frame_w because the prefix `╭─ NAME ─ ` was wider
-    /// than the entire chamber. Now the name itself gets
-    /// left-truncated to keep the header at most frame_w wide.
-    #[test]
-    fn banner_truncates_pathological_long_tool_name() {
-        let very_long = "MCP_TOOL:VERY_LONG_SERVER_NAME:VERY_LONG_FUNCTION_NAME";
-        let header = fit_banner_header(very_long, "/some/path", 40);
-        assert!(
-            header.as_str().width() <= 40,
-            "header must not exceed frame_w; got width {} for {:?}",
-            header.as_str().width(),
-            header,
-        );
-        assert!(header.starts_with("╭"));
-        assert!(header.ends_with("╮"));
-    }
-
-    // Partial assistant reply at abort time is preserved into the
-    // session with a trailer marking the interruption, so the LLM
-    // sees on the next turn what it had been saying. Mirrors
-    // opencode's `finalizeInterruptedAssistant` in
-    // `packages/opencode/src/session/prompt.ts`.
-    /// Phase 3 — abort with pending tool calls preserves them as
+    /// dirge-bfd: Ctrl-F search uses fuzzy matching (nucleo) — typos,
     /// structured entries on the stashed message. Pending entries
     /// stay Interrupted (no matching result arrived); on resume,
     /// `convert_history` will emit a [Tool execution was
@@ -6504,133 +5274,6 @@ mod tests {
             !content.contains("1 tool calls ran"),
             "leaked plural for singular case: {content:?}",
         );
-    }
-
-    /// `suggest_pattern` returns a literal placeholder for empty
-    /// input. The ask-dialog path that consumes it must detect the
-    /// placeholder and refuse to add it as an allowlist entry —
-    /// otherwise pressing "a" (allow always) on an empty invocation
-    /// would silently store `<edit this pattern>` as a real pattern.
-    /// The detection is exposed via `is_placeholder_pattern` so the
-    /// dialog code is unit-testable.
-    #[test]
-    fn placeholder_pattern_is_detectable() {
-        let p = suggest_pattern("bash", "");
-        assert!(
-            is_placeholder_pattern(&p),
-            "empty input should yield a detectable placeholder; got {p:?}",
-        );
-        let p = suggest_pattern("grep", "  \t  ");
-        assert!(is_placeholder_pattern(&p));
-        // A legit suggestion is NOT flagged as a placeholder.
-        let p = suggest_pattern("bash", "cargo test");
-        assert!(!is_placeholder_pattern(&p), "real pattern flagged: {p:?}");
-    }
-
-    // Whitespace-only or empty input must NOT collapse to a "* *"
-    // / "*" wildcard pattern that matches every subsequent call.
-    // The audit flagged this as a footgun: a user accidentally
-    // hitting "(a) allow always" on an empty bash invocation would
-    // permanently auto-allow ALL bash. Now we return a literal
-    // placeholder + the user has to type the pattern themselves.
-    #[test]
-    fn suggest_pattern_refuses_wildcard_on_empty_input() {
-        // Bash: empty / whitespace input should NOT yield "* *".
-        let p = suggest_pattern("bash", "");
-        assert_ne!(p, "* *", "empty bash input must not yield catch-all");
-        assert!(
-            !p.contains('*'),
-            "empty input should not contain wildcards: {p:?}"
-        );
-
-        let p = suggest_pattern("bash", "   \t  ");
-        assert_ne!(
-            p, "* *",
-            "whitespace-only bash input must not yield catch-all"
-        );
-        assert!(
-            !p.contains('*'),
-            "ws-only input should not contain wildcards: {p:?}"
-        );
-
-        // grep / find_files: same — empty must not yield "*"
-        let p = suggest_pattern("grep", "");
-        assert!(
-            !p.contains('*'),
-            "empty grep input must not yield wildcard: {p:?}"
-        );
-
-        // Unknown tool with empty input shouldn't yield catch-all.
-        let p = suggest_pattern("mcp_tool:foo", "");
-        assert!(!p.contains('*'), "unknown tool empty input: {p:?}");
-    }
-
-    // Non-empty inputs still produce the expected suggestion.
-    #[test]
-    fn suggest_pattern_works_for_non_empty_inputs() {
-        assert_eq!(suggest_pattern("bash", "cargo test --all"), "cargo *");
-        assert_eq!(suggest_pattern("grep", "fn foo bar"), "fn*");
-    }
-
-    /// User-reported bug: "allow always" on a write inside `src/`
-    /// stored `src/*` (single `*`, no slash-spanning), so the next
-    /// write under `src/agent/…` re-prompted. Maki's equivalent
-    /// (`maki-agent/src/permissions.rs:519`) uses `parent/**`. Pin
-    /// that the fix is in place for every path-shaped tool.
-    #[test]
-    fn suggest_pattern_path_tools_use_recursive_glob() {
-        assert_eq!(suggest_pattern("write", "src/main.rs"), "src/**");
-        assert_eq!(suggest_pattern("edit", "src/main.rs"), "src/**");
-        assert_eq!(
-            suggest_pattern("write", "src/agent/tools/foo.rs"),
-            "src/agent/tools/**"
-        );
-        assert_eq!(suggest_pattern("read", "src/main.rs"), "src/**");
-        assert_eq!(suggest_pattern("list_dir", "src/agent"), "src/**");
-        // Files at the repo root: `Path::parent` is "" — keep the
-        // existing `**` fallback so the rule is broad but explicit.
-        assert_eq!(suggest_pattern("write", "main.rs"), "**");
-    }
-
-    /// User-reported bug: `[a] allow always` on an MCP tool call
-    /// silently degraded to `allow once` because the catch-all
-    /// `_ => PLACEHOLDER` branch fired for `mcp_tool`. Result: the
-    /// permission allowlist never got an entry and every
-    /// subsequent call to the same MCP server re-prompted the
-    /// user.
-    #[test]
-    fn suggest_pattern_derives_server_wildcard_for_mcp_tool() {
-        let p = suggest_pattern("mcp_tool", "mcp_tool:lattice:lattice_expand");
-        assert_eq!(p, "mcp_tool:lattice:*");
-        // Multi-segment server names also work.
-        let p = suggest_pattern("mcp_tool", "mcp_tool:my-server:do_thing");
-        assert_eq!(p, "mcp_tool:my-server:*");
-    }
-
-    /// Malformed MCP input (missing colons, wrong umbrella) still
-    /// falls through to the placeholder rather than producing a
-    /// nonsense pattern.
-    #[test]
-    fn suggest_pattern_mcp_tool_malformed_input_uses_placeholder() {
-        assert!(is_placeholder_pattern(&suggest_pattern(
-            "mcp_tool", "garbage"
-        )));
-        assert!(is_placeholder_pattern(&suggest_pattern(
-            "mcp_tool",
-            "mcp_tool:"
-        )));
-        // Note: `"mcp_tool::"` parses with server="" so it
-        // correctly falls into the placeholder branch.
-        assert!(is_placeholder_pattern(&suggest_pattern(
-            "mcp_tool",
-            "mcp_tool::"
-        )));
-        // Note: `wrong_umbrella:lattice:foo` doesn't start with
-        // `mcp_tool:` so the umbrella check refuses.
-        assert!(is_placeholder_pattern(&suggest_pattern(
-            "mcp_tool",
-            "wrong:lattice:foo"
-        )));
     }
 
     // Rewind must sync tree.entries + message_store + leaf_id with
