+++ src/ui/tui/mod.rs
@@ -14,18 +14,9 @@
 //! alignment bugs caused by per-callsite column math become
 //! impossible.
 
-// Layout types are consumed by future migration phases (top frame
-// widget, chat region, side panels, bottom strip). The dead-code
-// warning will go away as those phases land.
-#[allow(dead_code)]
 pub mod bottom;
-#[allow(dead_code)]
 pub mod chat;
-#[allow(dead_code)]
 pub mod frame;
-#[allow(dead_code)]
 pub mod layout;
-#[allow(dead_code)]
 pub mod panels;
-#[allow(dead_code)]
 pub mod scene;
